Move circuit/service ID validation to functions
Moves circuit and service ID validation to dedicated functions in the
`admin::service::messages` module. These functions are made available
publicly and used elsewhere in libsplinter.

Also updates the error messages where these methods are used to include
the ID in the case that it's invalid.
